\name{find.time.var}
\alias{find.time.var}
\title{find time var}
\description{Determine which variable should be used for animation.}
\usage{find.time.var(p)}
\arguments{
  \item{p}{ggplot object.}
}

\value{An expression to be used for a time variable, or NULL if none
found.}

\author{Toby Dylan Hocking}





