\name{gganim-package}
\alias{gganim-package}
\docType{package}
\title{Convert ggplots to animations}
\description{The data in a ggplot is split using the time aesthetic
to form the frames of the animation.}
\details{
\tabular{ll}{Package: \tab gganim\cr
Maintainer: \tab Toby Dylan Hocking <toby.hocking@inria.fr>\cr
Author: \tab Toby Dylan Hocking\cr
Version: \tab 1.0\cr
License: \tab GPL-3\cr
Title: \tab Convert ggplots to animations\cr}
}
\author{Toby Dylan Hocking}

\keyword{ package }


